requirements/dev.txt
-r _base.txt

django-debug-toolbar
pytest-django
pytest